perf: 디렉터리 구조 재설정 및 공통 설정 export하여 테스트 병렬화
- 기존 코드는 테스트가 순차적으로 실행되어서 50초까지도 걸렸음
- 문제 3개 기준 20초 정도로 빠르게 실행 가능

[#145]
